refactor: 일급 컬렉션 (First Class Collection) 사용
Lotto 도메인을 domain 패키지로 따로 보관하고, List<List<Integer>> 대신 Lotto 클래스에 일급 컬렉션 적용
